refactor credentials service to main
ensure keytar is there and register service correctly

remove console log

remove extra line
