+++ src/vs/platform/credentials/common/credentials.ts
