+++ src/vs/platform/native/electron-main/nativeHostMainService.ts
@@ -42,11 +42,6 @@ export interface INativeHostMainService extends AddFirstParameterToFunctions<ICo
 
 export const INativeHostMainService = createDecorator<INativeHostMainService>('nativeHostMainService');
 
-interface ChunkedPassword {
-	content: string;
-	hasNextChunk: boolean;
-}
-
 export class NativeHostMainService extends Disposable implements INativeHostMainService {
 
 	declare readonly _serviceBrand: undefined;
@@ -822,123 +817,6 @@ export class NativeHostMainService extends Disposable implements INativeHostMain
 		} catch {
 			return undefined;
 		}
-	}
-
-	//#endregion
-
-
-	//#region Credentials
-
-	private static readonly MAX_PASSWORD_LENGTH = 2500;
-	private static readonly PASSWORD_CHUNK_SIZE = NativeHostMainService.MAX_PASSWORD_LENGTH - 100;
-
-	async getPassword(windowId: number | undefined, service: string, account: string): Promise<string | null> {
-		const keytar = await this.withKeytar();
-
-		const password = await keytar.getPassword(service, account);
-		if (password) {
-			try {
-				let { content, hasNextChunk }: ChunkedPassword = JSON.parse(password);
-				if (!content || !hasNextChunk) {
-					return password;
-				}
-
-				let index = 1;
-				while (hasNextChunk) {
-					const nextChunk = await keytar.getPassword(service, `${account}-${index}`);
-					const result: ChunkedPassword = JSON.parse(nextChunk!);
-					content += result.content;
-					hasNextChunk = result.hasNextChunk;
-					index++;
-				}
-
-				return content;
-			} catch {
-				return password;
-			}
-		}
-
-		return password;
-	}
-
-	async setPassword(windowId: number | undefined, service: string, account: string, password: string): Promise<void> {
-		const keytar = await this.withKeytar();
-		const MAX_SET_ATTEMPTS = 3;
-
-		// Sometimes Keytar has a problem talking to the keychain on the OS. To be more resilient, we retry a few times.
-		const setPasswordWithRetry = async (service: string, account: string, password: string) => {
-			let attempts = 0;
-			let error: any;
-			while (attempts < MAX_SET_ATTEMPTS) {
-				try {
-					await keytar.setPassword(service, account, password);
-					return;
-				} catch (e) {
-					error = e;
-					this.logService.warn('Error attempting to set a password: ', e);
-					attempts++;
-					await new Promise(resolve => setTimeout(resolve, 200));
-				}
-			}
-
-			// throw last error
-			throw error;
-		};
-
-		if (isWindows && password.length > NativeHostMainService.MAX_PASSWORD_LENGTH) {
-			let index = 0;
-			let chunk = 0;
-			let hasNextChunk = true;
-			while (hasNextChunk) {
-				const passwordChunk = password.substring(index, index + NativeHostMainService.PASSWORD_CHUNK_SIZE);
-				index += NativeHostMainService.PASSWORD_CHUNK_SIZE;
-				hasNextChunk = password.length - index > 0;
-
-				const content: ChunkedPassword = {
-					content: passwordChunk,
-					hasNextChunk: hasNextChunk
-				};
-
-				await setPasswordWithRetry(service, chunk ? `${account}-${chunk}` : account, JSON.stringify(content));
-				chunk++;
-			}
-
-		} else {
-			await setPasswordWithRetry(service, account, password);
-		}
-
-		this._onDidChangePassword.fire({ service, account });
-	}
-
-	async deletePassword(windowId: number | undefined, service: string, account: string): Promise<boolean> {
-		const keytar = await this.withKeytar();
-
-		const didDelete = await keytar.deletePassword(service, account);
-		if (didDelete) {
-			this._onDidChangePassword.fire({ service, account });
-		}
-
-		return didDelete;
-	}
-
-	async findPassword(windowId: number | undefined, service: string): Promise<string | null> {
-		const keytar = await this.withKeytar();
-
-		return keytar.findPassword(service);
-	}
-
-	async findCredentials(windowId: number | undefined, service: string): Promise<Array<{ account: string, password: string }>> {
-		const keytar = await this.withKeytar();
-
-		return keytar.findCredentials(service);
-	}
-
-	private async withKeytar(): Promise<typeof import('keytar')> {
-		if (this.environmentMainService.disableKeytar) {
-			throw new Error('keytar has been disabled via --disable-keytar option');
-		}
-
-		return await import('keytar');
 	}
 
 	//#endregion
